consolidated k6 tests all into one file
