+++ test/k6Tests.js
@@ -3,19 +3,99 @@
 /* eslint-disable func-names */
 /* eslint-disable no-unused-vars */
 /* eslint-disable import/no-unresolved */
-import { sleep, group } from 'k6';
-import TestGET from './k6Tests/loadTestGET.js';
-import TestPOST from './k6Tests/loadTestPOST.js';
-import TestREPORT from './k6Tests/loadTestREPORT.js';
-import TestHELPFUL from './k6Tests/loadTestHELPFUL.js';
+import http from 'k6/http';
+import { sleep, group, check } from 'k6';
+
+const testGET = () => {
+  const getRandomIntInclusive = (min, max) => Math.floor(Math.random() * (max - min + 1) + min);
+  const id = getRandomIntInclusive(1, 1000000);
+  const res = http.get(`http://localhost:8080/reviews/product/${id}?sort=relevance`);
+
+  sleep(1);
+
+  const checkRes = check(res, {
+    'status is 200 (there are reviews for this product id)': (r) => r.status === 200,
+    'status is 400 (there are no reviews for this product id)': (r) => r.status === 400,
+    'response body': (r) => r.body.length > 0 === true,
+    'sends valid data even if db request fails': (r) => r.body.includes('productName'),
+  });
+};
+
+const testPOST = () => {
+  const getRandomIntInclusive = (min, max) => Math.floor(Math.random() * (max - min + 1) + min);
+  const id = getRandomIntInclusive(1, 1000000);
+
+  const data = {
+    product_id: id,
+    rating: 5,
+    summary: 'k6test',
+    body: 'k6testtesttesttest',
+    recommend: true,
+    name: 'k6',
+    email: 'k6@k6.com',
+    characteristics: {
+      158622: 1,
+      158623: 1,
+      158624: 1,
+      158625: 1,
+    },
+    photos: [
+      'testImg1.jpg',
+      'testImg2.jpg',
+      'testImg3.jpg',
+      'testImg4.jpg',
+    ],
+  };
+
+  const params = {
+    headers: {
+      'Content-Type': 'application/json',
+    },
+  };
+
+  const res = http.post('http://localhost:8080/reviews/', JSON.stringify(data), params);
+
+  sleep(1);
+
+  const checkRes = check(res, {
+    'status is 201 (review posted successfully)': (r) => r.status === 201,
+    'status is 400 (could not post, review id already exists)': (r) => r.status === 400,
+  });
+};
+
+const testREPORT = () => {
+  const getRandomIntInclusive = (min, max) => Math.floor(Math.random() * (max - min + 1) + min);
+  const id = getRandomIntInclusive(1, 1000000);
+  const res = http.put(`http://localhost:8080/reviews/${id}/report`);
+
+  sleep(1);
+
+  const checkRes = check(res, {
+    'status is 204 (review reported)': (r) => r.status === 204,
+    'status is 400 (could not report review)': (r) => r.status === 400,
+  });
+};
+
+const testHELPFUL = () => {
+  const getRandomIntInclusive = (min, max) => Math.floor(Math.random() * (max - min + 1) + min);
+  const id = getRandomIntInclusive(1, 1000000);
+  const res = http.put(`http://localhost:8080/reviews/${id}/helpful`);
+
+  sleep(1);
+
+  const checkRes = check(res, {
+    'status is 204 (review marked helpful)': (r) => r.status === 204,
+    'status is 400 (could not mark review helpful)': (r) => r.status === 400,
+  });
+};
 
 export const options = {
   scenarios: {
     constant_request_rate: {
       executor: 'constant-arrival-rate',
-      rate: 100,
+      rate: 1000,
       timeUnit: '1s',
-      duration: '5s',
+      duration: '30s',
       preAllocatedVUs: 150,
       maxVUs: 225,
     },
@@ -25,19 +105,19 @@ export const options = {
 
 export default () => {
   group('GET request', () => {
-    TestGET();
+    testGET();
   });
 
   group('POST request', () => {
-    TestPOST();
+    testPOST();
   });
 
   group('report PUT request', () => {
-    TestREPORT();
+    testREPORT();
   });
 
   group('helpful PUT request', () => {
-    TestHELPFUL();
+    testHELPFUL();
   });
 
   sleep(1);
